+++ sdk/src/types/block/output/token_id.rs
@@ -13,33 +13,3 @@ impl From<FoundryId> for TokenId {
         Self::new(*foundry_id)
     }
 }
-
-#[allow(missing_docs)]
-pub mod dto {
-    use alloc::string::String;
-
-    use serde::{Deserialize, Serialize};
-
-    use super::*;
-    use crate::types::block::Error;
-
-    /// Describes a token id.
-    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-    pub struct TokenIdDto(pub String);
-
-    impl From<&TokenId> for TokenIdDto {
-        fn from(value: &TokenId) -> Self {
-            Self(prefix_hex::encode(**value))
-        }
-    }
-
-    impl TryFrom<&TokenIdDto> for TokenId {
-        type Error = Error;
-
-        fn try_from(value: &TokenIdDto) -> Result<Self, Self::Error> {
-            Ok(Self::new(
-                prefix_hex::decode(&value.0).map_err(|_e| Error::InvalidField("tokenId"))?,
-            ))
-        }
-    }
-}