+++ sdk/src/wallet/mod.rs
@@ -11,6 +11,8 @@ pub mod account;
 #[cfg(feature = "message_interface")]
 #[cfg_attr(docsrs, doc(cfg(feature = "message_interface")))]
 pub mod message_interface;
+#[cfg(any(feature = "stronghold", feature = "storage"))]
+pub(crate) mod migration;
 /// The wallet module.
 #[allow(clippy::module_inception)]
 pub mod wallet;