+++ sdk/CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `OutputMetadata::set_spent` method;
 - `ignore_if_bech32_mismatch` parameter to `Wallet::restore_backup()`;
 - `OutputWithMetadata::{into_output, into_metadata}` methods;
+- Storage and Backup migration;
 
 ### Changed
 
@@ -52,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client`, `Wallet`, and `Account` fns updated to reflect internal reorganization;
 - `AccountBuilder::new` now takes a wallet;
 - `InputSigningDataDto::chain` is now `Vec<u32>` instead of `Chain`;
+- Most `StrongholdAdapter` fns no longer require a mutable reference;
+- `StorageProvider` fns no longer require a mutable reference;
 - Custom `Serialize` and `Deserialize` impls for `WalletEvent` to have an integer `type` as tag;
 
 ### Removed
@@ -60,6 +63,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `client_without_tls` feature in favor of separate `client` and `tls` features;
 - `IncreaseNativeTokenSupplyOptions`;
 - `HARDENED` const;
+- `AliasIdDto`, `NftIdDto` and `TokenIdDto`;
+
+### Fixed
+
+- Storage records decryption;
 
 ## 0.3.0 - 2023-05-02